some working around, formatter still not working yet

+++ model/evstrm/mixer.go
@@ -29,7 +29,7 @@ func GetFirst(ev1, ev2 *model.LogEvent) bool {
 
 // GetEarliest returns whether ev1 has lowest timestamp rather than ev2
 func GetEarliest(ev1, ev2 *model.LogEvent) bool {
-	return uint64(ev1.GetTimestamp()) <= uint64(ev2.GetTimestamp())
+	return ev1.GetTimestamp() <= ev2.GetTimestamp()
 }
 
 func (m *Mixer) Reset(sel SelectF, i1, i2 Iterator) {
